Move CountTestResult to different folder
