+++ src/main/java/nl/ou/debm/test/SingleTestResultTest.java
@@ -1,5 +1,6 @@
 package nl.ou.debm.test;
 
+import nl.ou.debm.assessor.CountTestResult;
 import nl.ou.debm.assessor.ETestCategories;
 import nl.ou.debm.assessor.IAssessor;
 import nl.ou.debm.common.EArchitecture;
@@ -14,14 +15,14 @@ public class SingleTestResultTest
 {
     @Test
     void Basics(){
-        var s1 = new IAssessor.CountTestResult();
-        var s2 = new IAssessor.CountTestResult();
+        var s1 = new CountTestResult();
+        var s2 = new CountTestResult();
         assertEquals(s1,s2);
         s1.setActualValue(10); s1.setActualValue(10);
         assertEquals(s1,s2);
-        var s3 = new IAssessor.CountTestResult(ETestCategories.FEATURE1_AGGREGATED, EArchitecture.X64ARCH, ECompiler.CLANG, EOptimize.OPTIMIZE,
+        var s3 = new CountTestResult(ETestCategories.FEATURE1_AGGREGATED, EArchitecture.X64ARCH, ECompiler.CLANG, EOptimize.OPTIMIZE,
                 0, 10, 15);
-        var s4 = new IAssessor.CountTestResult(ETestCategories.FEATURE1_AGGREGATED, EArchitecture.X64ARCH, ECompiler.CLANG, EOptimize.OPTIMIZE,
+        var s4 = new CountTestResult(ETestCategories.FEATURE1_AGGREGATED, EArchitecture.X64ARCH, ECompiler.CLANG, EOptimize.OPTIMIZE,
                 0, 10, 15);
         assertEquals(s3,s4);
         assertNotEquals(s1,s4);
